EditProfileModelと合併
